requirements.txt
pydantic==2.11.7
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.32.3
openai>=1.30.0
